fix: 리뷰 내용 수정
