+++ be/market/src/main/java/com/carrot/market/chatroom/infrastructure/ChatroomRepository.java
@@ -1,5 +1,7 @@
 package com.carrot.market.chatroom.infrastructure;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -7,4 +9,5 @@ import com.carrot.market.chatroom.domain.Chatroom;
 
 @Repository
 public interface ChatroomRepository extends JpaRepository<Chatroom, Long> {
+	Optional<Chatroom> findByProductIdAndPurchaserId(Long productId, Long purchaserId);
 }
