+++ be/market/src/main/java/com/carrot/market/global/config/kafka/ProducerConfiguration.java
@@ -19,7 +19,6 @@ import lombok.RequiredArgsConstructor;
 
 @EnableKafka
 @Configuration
-@RequiredArgsConstructor
 public class ProducerConfiguration {
 
 	@Bean
