+++ be/market/src/main/java/com/carrot/market/global/util/KafkaConstant.java
@@ -1,7 +1,11 @@
 package com.carrot.market.global.util;
 
-public abstract class KafkaConstant {
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class KafkaConstant {
 	public static final String KAFKA_TOPIC = "bee-chat";
 	public static final String GROUP_ID = "foo";
 	public static final String KAFKA_BROKER = "localhost:9092";
+
 }
